Optimize game object spawning and visibility, modify game balancing
Streamlined the way game objects such as cells and asteroids spawn within the visible area, by focusing the code to spawn objects on the GameManager and cleaning up irrelevant object creation in other classes. This includes the refactoring of calculating visible area bounds to improve game performance, by reducing unnecessary function calls. The code also modifies game balancing, by adjusting the number of blobs a player starts with from 1 to 3. This change was implemented to enhance the early game experience for new players. Implemented changes are still a work in progress: further optimizations for spawning cells in and around the visible area are planned.